requirements.txt
Flask==3.0.0
flask-cors==3.0.10
numpy>=1.21.2
tensorflow-cpu==2.14.0
Pillow== 10.1.0